refactor(storage): split typed storage headers
