+++ framework/core/src/libyijinjing/include/kungfu/yijinjing/storage/episode_manifest.h
@@ -4,16 +4,11 @@
 #define KUNGFU_YIJINJING_STORAGE_EPISODE_MANIFEST_H
 
 #include <cstdint>
-#include <functional>
-#include <map>
-#include <optional>
 #include <string>
-#include <variant>
-#include <vector>
 
 #include <nlohmann/json.hpp>
 
-#include <kungfu/yijinjing/schema/types.h>
+#include <kungfu/yijinjing/storage/episode_manifest_types.h>
 
 namespace kungfu::yijinjing::storage {
 
@@ -22,10 +17,6 @@ namespace kungfu::yijinjing::storage {
 // deterministic typed fold is the canonical in-memory derivation; JSON is an
 // edge projection only (CLI, export, binding return values). Fold semantics
 // are specified in docs/episode-manifest-trust-boundary.md.
-inline constexpr const char *EPISODE_MANIFEST_SCHEMA_V1 = "kungfu.episode.manifest/v1";
-inline constexpr const char *EPISODE_MANIFEST_NAMESPACE = "storage";
-inline constexpr const char *EPISODE_MANIFEST_NAME = "episode-manifest";
-
 // Edge-level input options. These carry std::string for ergonomic callers; the
 // store copies them into fixed-layout POD journal records. They are not the
 // stored record and never become the fact substrate.
@@ -102,132 +93,9 @@ struct episode_recover_options {
 // with manifest_writer_busy; see docs/episode-manifest-trust-boundary.md §3.
 [[nodiscard]] std::string episode_manifest_writer_lock_path(const std::string &runtime_dir);
 
-// A manifest frame whose carrier type is not a v1 record, or whose
-// schema_version is newer than this reader understands. It is preserved in
-// the typed stream with provenance but never folded into an Episode view, so
-// a newer writer does not brick an older reader.
-struct episode_manifest_unknown_record {
-  int32_t carrier_type = 0;
-  uint32_t schema_version = 0; // 0 when not extractable from the frame
-  bool unknown_version = false;
-};
-
-// One manifest journal frame decoded into its typed POD record, in append
-// order, with the manifest frame provenance the edge projections expose.
-struct episode_manifest_record {
-  uint64_t manifest_frame_uid = 0;
-  int64_t manifest_gen_time = 0;
-  std::variant<episode_manifest_unknown_record, yijinjing::types::EpisodeOpen, yijinjing::types::EpisodeHeartbeat,
-               yijinjing::types::EpisodeFrameAttached, yijinjing::types::EpisodeRefAttached,
-               yijinjing::types::EpisodeClosed, yijinjing::types::EpisodeRootCommitted>
-      body = episode_manifest_unknown_record{};
-};
-
-using episode_manifest_record_visitor = std::function<void(const episode_manifest_record &)>;
-
-// The typed current view of one Episode, derived by the deterministic fold.
-// Identity fields come from the first EpisodeOpen; watermark fields are
-// last-writer-wins in append order; collections keep append order including
-// duplicates (the journal is the authority for what was claimed).
-struct episode_current_view {
-  uint64_t episode_id = 0;
-  bool opened = false;
-  bool closed = false;
-  size_t open_count = 0;
-  size_t close_count = 0;
-  yijinjing::types::EpisodeOpen open = {};
-  uint64_t open_manifest_frame_uid = 0;
-  int64_t open_manifest_gen_time = 0;
-  bool heartbeat_seen = false;
-  int64_t update_time = 0;
-  uint64_t claimed_frame_count = 0;
-  bool last_frame_uid_seen = false;
-  uint64_t last_frame_uid = 0;
-  size_t unique_frame_count = 0;
-  yijinjing::types::EpisodeClosed close = {};
-  std::vector<yijinjing::enums::EpisodeStatus> close_statuses = {};
-  // ADR-0043: the recorded content-root claim. First-committed-wins (a root
-  // names a sealed identity once); later duplicates are fsck diagnostics.
-  bool root_seen = false;
-  size_t root_count = 0;
-  yijinjing::types::EpisodeRootCommitted root = {};
-  std::vector<episode_manifest_record> records = {};
-  std::vector<size_t> frame_indices = {};
-  std::vector<size_t> ref_indices = {};
-  std::vector<uint64_t> duplicate_frame_uids = {};
-  size_t missing_frame_uid_count = 0;
-
-  [[nodiscard]] const yijinjing::types::EpisodeFrameAttached &frame_at(size_t position) const {
-    return std::get<yijinjing::types::EpisodeFrameAttached>(records[frame_indices[position]].body);
-  }
-
-  [[nodiscard]] const yijinjing::types::EpisodeRefAttached &ref_at(size_t position) const {
-    return std::get<yijinjing::types::EpisodeRefAttached>(records[ref_indices[position]].body);
-  }
-};
-
-struct episode_manifest_fold {
-  std::map<uint64_t, episode_current_view> episodes = {};
-  size_t total_record_count = 0;
-  size_t unknown_record_count = 0;
-  size_t unfolded_record_count = 0;
-};
-
-struct episode_fsck_issue {
-  std::string code = {};
-  std::optional<uint64_t> episode_id = {};
-  std::optional<uint64_t> dependency_episode_id = {};
-  std::optional<uint64_t> frame_uid = {};
-  std::optional<uint64_t> dependent_frame_uid = {};
-  std::optional<uint64_t> count = {};
-  std::optional<int32_t> status = {};
-  std::optional<uint64_t> claimed = {};
-  std::optional<uint64_t> actual = {};
-  std::optional<uint64_t> recorded_covered_record_count = {};
-  std::optional<uint64_t> computed_covered_record_count = {};
-  std::optional<std::string> role = {};
-  std::optional<std::string> ref_id = {};
-  std::optional<std::string> ref_hash = {};
-  std::optional<std::string> detail = {};
-  std::optional<std::string> reason = {};
-  std::optional<std::string> algorithm = {};
-  std::optional<std::string> recorded = {};
-  std::optional<std::string> computed = {};
-};
-
-struct episode_fsck_result {
-  bool ok = true;
-  std::string status = "ok";
-  std::string schema = EPISODE_MANIFEST_SCHEMA_V1;
-  std::string runtime_dir = {};
-  std::string authority = "yijinjing-journal";
-  bool degraded = false;
-  std::vector<episode_fsck_issue> errors = {};
-  std::vector<episode_fsck_issue> warnings = {};
-  uint64_t episode_manifest_records = 0;
-  uint64_t episodes = 0;
-  uint64_t unknown_records = 0;
-  uint64_t unfolded_records = 0;
-  std::optional<episode_current_view> episode = {};
-};
-
 [[nodiscard]] nlohmann::json render_episode_fsck_issue(const episode_fsck_issue &issue);
 
 [[nodiscard]] nlohmann::json render_episode_fsck_result(const episode_fsck_result &result);
-
-// ADR-0043: the content root of one Episode's owned claim sequence — a linear
-// hash chain over the covered records in manifest append order (the first
-// EpisodeOpen, every EpisodeFrameAttached / EpisodeRefAttached, and the first
-// terminal EpisodeClosed). Heartbeats, manifest provenance, unknown records,
-// and the root record itself stay outside identity. Each record contributes
-// its hana field bytes in declaration order (padding never enters the hash).
-struct episode_content_root {
-  uint32_t covered_record_count = 0;
-  std::string algorithm = {};
-  std::string value = {};
-};
-
-[[nodiscard]] episode_content_root compute_episode_content_root(const episode_current_view &view);
 
 class content_store;
