+++ framework/core/src/libyijinjing/include/kungfu/yijinjing/storage/manifest_catalog.h
@@ -3,14 +3,13 @@
 #ifndef KUNGFU_YIJINJING_STORAGE_MANIFEST_CATALOG_H
 #define KUNGFU_YIJINJING_STORAGE_MANIFEST_CATALOG_H
 
-#include <optional>
 #include <string>
 #include <vector>
 
 #include <nlohmann/json.hpp>
 
-#include <kungfu/yijinjing/schema/types.h>
 #include <kungfu/yijinjing/storage/content_store.h>
+#include <kungfu/yijinjing/storage/manifest_catalog_types.h>
 
 namespace kungfu::yijinjing::storage {
 
@@ -24,12 +23,6 @@ namespace kungfu::yijinjing::storage {
 // committed by content hash into the content store so the JSON edge and the
 // cross-store sync root stay byte-reproducible. JSON is an edge projection
 // only, never the contract.
-inline constexpr const char *MANIFEST_CATALOG_SCHEMA_V1 = "kungfu.storage.manifest-catalog/v1";
-inline constexpr const char *MANIFEST_CATALOG_NAMESPACE = "storage";
-inline constexpr const char *MANIFEST_CATALOG_NAME = "manifest-catalog";
-// Content-store namespace holding the canonical accepted entries documents.
-inline constexpr const char *MANIFEST_ENTRIES_CONTENT_NAMESPACE = "manifests";
-
 // JSON edge schemas. These name edge projections and exchange documents over
 // the journal records, not stored contracts.
 inline constexpr const char *STORAGE_SOURCE_RECORD_SCHEMA_V1 = "kungfu.storage.source-record/v1";
@@ -70,50 +63,6 @@ struct storage_issue {
 [[nodiscard]] nlohmann::json build_storage_payload_inventory(const nlohmann::json &entries);
 
 [[nodiscard]] nlohmann::json build_storage_schema_inventory(const nlohmann::json &entries);
-
-// Typed POD records folded off the manifest-catalog journal, in append order.
-struct manifest_catalog_journal_records {
-  std::vector<yijinjing::types::ImportManifestAccepted> manifests = {};
-  std::vector<yijinjing::types::ManifestEntryRecorded> entries = {};
-  std::vector<yijinjing::types::ExportBundleRecorded> exports = {};
-  std::vector<yijinjing::types::ChannelCursorUpdated> cursors = {};
-};
-
-struct manifest_catalog_fsck_issue {
-  std::string code = {};
-  std::optional<std::string> source_id = {};
-  std::optional<std::string> manifest_id = {};
-  std::optional<std::string> error = {};
-  std::optional<std::string> subject = {};
-  std::optional<std::string> payload_hash = {};
-  std::optional<std::string> state = {};
-  std::optional<std::string> kind = {};
-  std::optional<std::string> entry_source_id = {};
-  std::optional<uint64_t> manifest_uid = {};
-  std::optional<uint64_t> entry_index = {};
-  std::optional<uint64_t> expected = {};
-  std::optional<uint64_t> actual = {};
-  std::optional<std::string> expected_text = {};
-  std::optional<std::string> actual_text = {};
-  std::optional<bool> intentional = {};
-};
-
-struct manifest_catalog_fsck_result {
-  bool ok = true;
-  std::string status = "ok";
-  bool degraded = false;
-  std::string schema = MANIFEST_CATALOG_SCHEMA_V1;
-  std::string runtime_dir = {};
-  std::string authority = "yijinjing-journal";
-  std::vector<manifest_catalog_fsck_issue> errors = {};
-  std::vector<manifest_catalog_fsck_issue> warnings = {};
-  uint64_t manifests = 0;
-  uint64_t manifest_entries = 0;
-  uint64_t payloads = 0;
-  uint64_t entries_documents = 0;
-  uint64_t exports = 0;
-  uint64_t cursors = 0;
-};
 
 class manifest_catalog_store {
 public:
